+++ .gitignore
@@ -1,2 +1,2 @@
-firebase.config.json
+firebase_manager/firebase.config.json
 .env
