+++ models/server_passwords.py
@@ -1,4 +1,3 @@
-import json
 import logging
 from typing import List
 from firebase_admin import db
