+++ models/server_users.py
@@ -1,4 +1,3 @@
-import json
 import logging
 from typing import List
